Update jQuery ScrollTo to v2.1.1.
