+++ README.rst
@@ -22,7 +22,7 @@ Starring
 * `jQuery UI (1.11.4) <https://github.com/jquery/jquery-ui>`_
 * `jQuery DataTables (1.10.7) <https://github.com/DataTables/DataTables>`_
 * `jQuery Dynamic Formset (1.2) <https://github.com/elo80ka/django-dynamic-formset>`_
-* `jQuery ScrollTo (2.1.0) <https://github.com/flesler/jquery.scrollTo>`_
+* `jQuery ScrollTo (2.1.1) <https://github.com/flesler/jquery.scrollTo>`_
 * `jQuery Smooth Scroll (1.5.5) <https://github.com/kswedberg/jquery-smooth-scroll>`_
 * `Twitter Bootstrap (3.3.4) <https://github.com/twbs/bootstrap>`_
 * `iOS-Orientationchange-Fix <https://github.com/scottjehl/iOS-Orientationchange-Fix>`_
