+++ djfrontend/templatetags/settings.py
@@ -8,6 +8,6 @@ DJFRONTEND_JQUERY_DEFAULT = '1.11.3'
 DJFRONTEND_JQUERYUI_DEFAULT = '1.11.4'
 DJFRONTEND_JQUERY_DATATABLES_VERSION_DEFAULT = '1.10.7'
 DJFRONTEND_JQUERY_FORMSET_DEFAULT = '1.2'
-DJFRONTEND_JQUERY_SCROLLTO_DEFAULT = '2.1.0'
+DJFRONTEND_JQUERY_SCROLLTO_DEFAULT = '2.1.1'
 DJFRONTEND_JQUERY_SMOOTHSCROLL_DEFAULT = '1.5.5'
 DJFRONTEND_TWBS_VERSION_DEFAULT = '3.3.4'
