+++ docs/source/granularity.rst
@@ -32,7 +32,7 @@ If you use an optional value other than the ones listed, you will have to add th
 +-----------------------------------+---------------------------------------+---------------+-------------------+
 |djfrontend_jquery_formset          |DJFRONTEND_JQUERY_FORMSET              |1.2            |                   |
 +-----------------------------------+---------------------------------------+---------------+-------------------+
-|djfrontend_jquery_scrollto         |DJFRONTEND_JQUERY_SCROLLTO             |2.1.0          |                   |
+|djfrontend_jquery_scrollto         |DJFRONTEND_JQUERY_SCROLLTO             |2.1.1          |                   |
 +-----------------------------------+---------------------------------------+---------------+-------------------+
 |djfrontend_jquery_smoothscroll     |DJFRONTEND_JQUERY_SMOOTHSCROLL         |1.5.5          |                   |
 +-----------------------------------+---------------------------------------+---------------+-------------------+
